+++ CHANGES.md
@@ -1,3 +1,7 @@
+#### Version 0.0.3 - 2017/10/16
+* Making links actually clickable, using JS to disable
+* Adding minified versions of CSS and JS files
+
 #### Version 0.0.2 - 2016/01/10
 * general code cleanup